Optimized and refactored jdbc tests.
