+++ eventuate-client-java-micronaut-jdbc/build.gradle
@@ -18,6 +18,8 @@ dependencies {
     compile project(":eventuate-client-java-micronaut")
     compile project(":eventuate-client-java-jdbc-common-tests")
 
+    compile "io.eventuate.common:eventuate-common-micronaut-in-memory-database:$eventuateCommonVersion"
+
     compile "io.eventuate.common:eventuate-common-micronaut-spring-jdbc:$eventuateCommonVersion"
 
     compile "io.micronaut:micronaut-spring"
