+++ eventuate-client-java-spring-jdbc/src/main/java/io/eventuate/javaclient/spring/jdbc/EmbeddedTestAggregateStoreConfiguration.java
@@ -3,6 +3,7 @@ package io.eventuate.javaclient.spring.jdbc;
 import io.eventuate.common.inmemorydatabase.EventuateDatabaseScriptSupplier;
 import io.eventuate.common.jdbc.EventuateCommonJdbcOperations;
 import io.eventuate.common.jdbc.EventuateJdbcStatementExecutor;
+import io.eventuate.common.jdbc.EventuateSchema;
 import io.eventuate.common.jdbc.EventuateTransactionTemplate;
 import io.eventuate.common.spring.inmemorydatabase.EventuateCommonInMemoryDatabaseConfiguration;
 import io.eventuate.common.spring.jdbc.EventuateCommonJdbcOperationsConfiguration;
@@ -37,8 +38,21 @@ public class EmbeddedTestAggregateStoreConfiguration {
   }
 
   @Bean
-  public EventuateJdbcAccess eventuateJdbcAccess(EventuateTransactionTemplate eventuateTransactionTemplate, EventuateJdbcStatementExecutor eventuateJdbcStatementExecutor, EventuateCommonJdbcOperations eventuateCommonJdbcOperations) {
-    return new EventuateJdbcAccessImpl(eventuateTransactionTemplate, eventuateJdbcStatementExecutor, eventuateCommonJdbcOperations);
+  public TransactionTemplate transactionTemplate(DataSource dataSource) {
+    return new TransactionTemplate(new DataSourceTransactionManager(dataSource));
+  }
+
+
+  @Bean
+  public EventuateJdbcAccess eventuateJdbcAccess(EventuateTransactionTemplate eventuateTransactionTemplate,
+                                                 EventuateJdbcStatementExecutor eventuateJdbcStatementExecutor,
+                                                 EventuateCommonJdbcOperations eventuateCommonJdbcOperations,
+                                                 EventuateSchema eventuateSchema) {
+
+    return new EventuateJdbcAccessImpl(eventuateTransactionTemplate,
+            eventuateJdbcStatementExecutor,
+            eventuateCommonJdbcOperations,
+            eventuateSchema);
   }
 
   @Bean
@@ -59,10 +73,5 @@ public class EmbeddedTestAggregateStoreConfiguration {
   @Bean
   public EventuateClientScheduler eventHandlerRecoveryScheduler() {
     return new JdkTimerBasedEventuateClientScheduler();
-  }
-
-  @Bean
-  public TransactionTemplate transactionTemplate(DataSource dataSource) {
-    return new TransactionTemplate(new DataSourceTransactionManager(dataSource));
   }
 }
