+++ eventuate-client-java-spring-jdbc/src/test/java/io/eventuate/javaclient/spring/jdbc/CustomEventuateJdbcAccessImplTest.java
@@ -3,34 +3,27 @@ package io.eventuate.javaclient.spring.jdbc;
 import io.eventuate.common.jdbc.EventuateSchema;
 import org.junit.Before;
 import org.junit.runner.RunWith;
-import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.boot.test.IntegrationTest;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Scope;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import javax.sql.DataSource;
 import java.util.List;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = {CustomEventuateJdbcAccessImplTest.Config.class, EventuateJdbcAccessImplTest.Config.class})
+@SpringApplicationConfiguration(classes = {EmbeddedTestAggregateStoreConfiguration.class, CustomEventuateJdbcAccessImplTest.Config.class})
 @IntegrationTest
 public class CustomEventuateJdbcAccessImplTest extends EventuateJdbcAccessImplTest {
 
   public static class Config {
-
     @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public DataSource dataSource() {
-      EmbeddedDatabaseBuilder builder = new EmbeddedDatabaseBuilder();
-      return builder.setType(EmbeddedDatabaseType.H2).build();
+    public JdbcTemplate jdbcTemplate(DataSource dataSource) {
+      return new JdbcTemplate(dataSource);
     }
 
     @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     public EventuateSchema eventuateSchema() {
       return new EventuateSchema("custom");
     }
@@ -56,5 +49,7 @@ public class CustomEventuateJdbcAccessImplTest extends EventuateJdbcAccessImplTe
     List<String> lines = loadSqlScriptAsListOfLines("eventuate-embedded-schema.sql");
     for (int i = 0; i < 2; i++) lines.set(i, lines.get(i).replace("eventuate", "custom"));
     executeSql(lines);
+
+    clear();
   }
 }
