+++ eventuate-client-java-spring-jdbc/src/test/java/io/eventuate/javaclient/spring/jdbc/DefaultEventuateJdbcAccessImplTest.java
@@ -1,37 +1,36 @@
 package io.eventuate.javaclient.spring.jdbc;
 
 import io.eventuate.common.jdbc.EventuateSchema;
+import org.junit.Before;
 import org.junit.runner.RunWith;
-import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.boot.test.IntegrationTest;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Scope;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import javax.sql.DataSource;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = {DefaultEventuateJdbcAccessImplTest.Config.class, EventuateJdbcAccessImplTest.Config.class})
+@SpringApplicationConfiguration(classes = {EmbeddedTestAggregateStoreConfiguration.class, DefaultEventuateJdbcAccessImplTest.Config.class})
 @IntegrationTest
 public class DefaultEventuateJdbcAccessImplTest extends EventuateJdbcAccessImplTest {
 
   public static class Config {
-
     @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public DataSource dataSource() {
-      EmbeddedDatabaseBuilder builder = new EmbeddedDatabaseBuilder();
-      return builder.setType(EmbeddedDatabaseType.H2).addScript("eventuate-embedded-schema.sql").build();
+    public JdbcTemplate jdbcTemplate(DataSource dataSource) {
+      return new JdbcTemplate(dataSource);
     }
 
     @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     public EventuateSchema eventuateSchema() {
       return new EventuateSchema();
     }
+  }
+
+  @Before
+  public void init() {
+    clear();
   }
 
   @Override
