+++ eventuate-client-java-spring-jdbc/src/test/java/io/eventuate/javaclient/spring/jdbc/EventuateJdbcAccessImplTest.java
@@ -1,81 +1,12 @@
 package io.eventuate.javaclient.spring.jdbc;
 
-import io.eventuate.common.common.spring.jdbc.EventuateSpringJdbcStatementExecutor;
-import io.eventuate.common.common.spring.jdbc.EventuateSpringTransactionTemplate;
 import io.eventuate.common.jdbc.*;
-import io.eventuate.common.jdbc.sqldialect.SqlDialectSelector;
-import io.eventuate.common.spring.jdbc.sqldialect.SqlDialectConfiguration;
 import io.eventuate.javaclient.jdbc.EventuateJdbcAccess;
-import io.eventuate.javaclient.jdbc.EventuateJdbcAccessImpl;
 import io.eventuate.javaclient.jdbc.common.tests.CommonEventuateJdbcAccessImplTest;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.beans.factory.config.ConfigurableBeanFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.Scope;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.support.TransactionTemplate;
-
-import javax.sql.DataSource;
 
 public abstract class EventuateJdbcAccessImplTest extends CommonEventuateJdbcAccessImplTest {
-
-  @Configuration
-  @Import(SqlDialectConfiguration.class)
-  public static class Config {
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public EventuateTransactionTemplate eventuateTransactionTemplate(TransactionTemplate transactionTemplate) {
-      return new EventuateSpringTransactionTemplate(transactionTemplate);
-    }
-
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public EventuateJdbcStatementExecutor eventuateJdbcStatementExecutor(JdbcTemplate jdbcTemplate) {
-      return new EventuateSpringJdbcStatementExecutor(jdbcTemplate);
-    }
-
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public EventuateCommonJdbcOperations eventuateCommonJdbcOperations(EventuateJdbcStatementExecutor eventuateJdbcStatementExecutor,
-                                                                       SqlDialectSelector sqlDialectSelector,
-                                                                       @Value("${spring.datasource.driver-class-name}") String driver) {
-
-      return new EventuateCommonJdbcOperations(eventuateJdbcStatementExecutor, sqlDialectSelector.getDialect(driver));
-    }
-
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public EventuateJdbcAccess eventuateJdbcAccess(EventuateTransactionTemplate eventuateTransactionTemplate, EventuateJdbcStatementExecutor eventuateJdbcStatementExecutor, EventuateCommonJdbcOperations eventuateCommonJdbcOperations) {
-      return new EventuateJdbcAccessImpl(eventuateTransactionTemplate, eventuateJdbcStatementExecutor, eventuateCommonJdbcOperations);
-    }
-
-
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public JdbcTemplate jdbcTemplate(DataSource dataSource) {
-      return new JdbcTemplate(dataSource);
-    }
-
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public EventuateJdbcAccess eventuateJdbcAccess(EventuateTransactionTemplate eventuateTransactionTemplate,
-                                                   EventuateJdbcStatementExecutor eventuateJdbcStatementExecutor,
-                                                   EventuateCommonJdbcOperations eventuateCommonJdbcOperations,
-                                                   EventuateSchema eventuateSchema) {
-      return new EventuateJdbcAccessImpl(eventuateTransactionTemplate, eventuateJdbcStatementExecutor, eventuateCommonJdbcOperations, eventuateSchema);
-    }
-
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public TransactionTemplate transactionTemplate(DataSource dataSource) {
-      return new TransactionTemplate(new DataSourceTransactionManager(dataSource));
-    }
-  }
 
   @Autowired
   private EventuateJdbcStatementExecutor eventuateJdbcStatementExecutor;
@@ -83,6 +14,13 @@ public abstract class EventuateJdbcAccessImplTest extends CommonEventuateJdbcAcc
   @Autowired
   private EventuateJdbcAccess eventuateJdbcAccess;
 
+  @Autowired
+  private EventuateSchema eventuateSchema;
+
+  @Override
+  protected EventuateSchema getEventuateSchema() {
+    return eventuateSchema;
+  }
 
   @Test
   @Override
